Struct alloy_pubsub::frontend::PubSubFrontend

pub struct PubSubFrontend {
    tx: UnboundedSender<PubSubInstruction>,
    channel_size: usize,
}

A PubSubFrontend is Transport composed of a channel to a running PubSub service.

Fields
tx: UnboundedSender<PubSubInstruction>channel_size: usize

The number of items to buffer in new subscription channels. Defaults to 16. See [tokio::sync::broadcast::channel] for a description.

Implementations

impl PubSubFrontend

pub(crate) const fn new(tx: UnboundedSender<PubSubInstruction>) -> Self

Create a new frontend.

pub fn get_subscription( &self, id: U256 ) -> impl Future<Output = TransportResult<RawSubscription>> + Send + 'static

Get the subscription ID for a local ID.

pub fn unsubscribe(&self, id: U256) -> TransportResult<()>

Unsubscribe from a subscription.

pub fn send( &self, req: SerializedRequest ) -> impl Future<Output = TransportResult<Response>> + Send + 'static

Send a request.

pub fn send_packet(&self, req: RequestPacket) -> TransportFut<'static>

Send a packet of requests, by breaking it up into individual requests.

Once all responses are received, we return a single response packet.

pub const fn channel_size(&self) -> usize

Get the currently configured channel size. This is the number of items to buffer in new subscription channels. Defaults to 16. See [tokio::sync::broadcast] for a description of relevant behavior.

pub fn set_channel_size(&mut self, channel_size: usize)

Set the channel size. This is the number of items to buffer in new subscription channels. Defaults to 16. See [tokio::sync::broadcast] for a description of relevant behavior.