Struct alloy_pubsub::managers::sub::SubscriptionManager

pub(crate) struct SubscriptionManager {
    local_to_sub: BiBTreeMap<B256, ActiveSubscription>,
    local_to_server: BiBTreeMap<B256, U256>,
}

Fields
local_to_sub: BiBTreeMap<B256, ActiveSubscription>

The subscriptions.

local_to_server: BiBTreeMap<B256, U256>

Tracks the CURRENT server id for a subscription.

Implementations

impl SubscriptionManager

pub(crate) fn iter(&self) -> impl Iterator<Item = (&B256, &ActiveSubscription)>

Get an iterator over the subscriptions.

pub(crate) fn len(&self) -> usize

Get the number of subscriptions.

fn insert( &mut self, request: SerializedRequest, server_id: U256, channel_size: usize ) -> RawSubscription

Insert a subscription.

pub(crate) fn upsert( &mut self, request: SerializedRequest, server_id: U256, channel_size: usize ) -> RawSubscription

Insert or update the server_id for a subscription.

pub(crate) fn local_id_for(&self, server_id: U256) -> Option<B256>

De-alias an alias, getting the original ID.

pub(crate) fn drop_server_ids(&mut self)

Drop all server_ids.

fn change_server_id(&mut self, local_id: B256, server_id: U256)

Change the server_id of a subscription.

pub(crate) fn remove_sub(&mut self, local_id: B256)

Remove a subscription by its local_id.

pub(crate) fn notify(&mut self, notification: EthNotification)

Notify the subscription channel of a new value, if the sub is known, and if any receiver exists. If the sub id is unknown, or no receiver exists, the notification is dropped.

pub(crate) fn get_subscription(&self, local_id: B256) -> Option<RawSubscription>

Get a receiver for a subscription.